Converted a number of ints to size_t in SAPT code
Should prevent integer overflow. May not be exhaustive though.

+++ psi4/src/psi4/libsapt_solver/amplitudes.cc
@@ -620,7 +620,7 @@ void SAPT2::ijkl_to_ikjl(double *tARAR, int ilength, int jlength, int klength,
 
   for(int i=0; i<ilength; i++) {
     for(int l=0; l<llength; l++) {
-      long int ijkl = i*jlength*klength*llength + l;
+      long int ijkl = (long int)i*jlength*klength*llength + l;
       C_DCOPY(jlength*klength,&(tARAR[ijkl]),llength,X,1);
       for(int j=0; j<jlength; j++) {
         for(int k=0; k<klength; k++) {
