+++ psi4/src/psi4/libsapt_solver/sapt0.h
@@ -179,11 +179,11 @@ struct SAPTDFInts {
   bool dress_disk_;
   bool active_;
 
-  int i_length_;
-  int j_length_;
-  int ij_length_;
-  int i_start_;
-  int j_start_;
+  size_t i_length_;
+  size_t j_length_;
+  size_t ij_length_;
+  size_t i_start_;
+  size_t j_start_;
 
   double **B_p_;
   double **B_d_;
@@ -206,10 +206,10 @@ struct SAPTDFInts {
 
 struct Iterator {
 
-  int num_blocks;
+  size_t num_blocks;
   int* block_size;
 
-  int curr_block;
+  size_t curr_block;
   long int curr_size;
 
   ~Iterator() { free(block_size); };
